refactor(gateway): extract proxy/text/media mixins from base.py (shard s1)
